A bit of cleaning before PR
