+++ src/compiler/ir/mod.rs
@@ -1,4 +1,3 @@
-use self::ir_nodes::func::IRFunction;
 pub mod ir_nodes;
 pub mod print;
 pub mod analysis;
@@ -6,7 +5,8 @@ pub mod builder;
 pub mod context;
 pub mod errors;
 pub mod transform;
+pub mod data_type;
 #[derive(Default)]
 pub struct IRProgram {
-    pub functions:Vec<IRFunction>
+    pub functions:Vec<ir_nodes::IRFunction>
 }
