+++ src/compiler/mod.rs
@@ -1,0 +1,3 @@
+pub mod ast;
+pub mod ir;
+pub mod parser;
