--- src/ir/ir_nodes/block.rs
@@ -1,8 +1,0 @@
-use super::{variable::IRVariable, stmt::IRStatement};
-
-pub struct IRBlock {
-    variables: Vec<IRVariable>,
-    scope: usize,
-    function_id: usize, 
-    stmt: Vec<IRStatement>,
-}
