--- src/ir/ir_nodes/expr.rs
@@ -1,8 +1,0 @@
-use super::{variable::Value, binop::IRBinaryOp, call::IRCall, unaryop::IRUnaryOp};
-pub enum IRExpression {
-    BinaryOp(IRBinaryOp),
-    UnaryOp(IRUnaryOp),
-    Call(IRCall),
-    Value(Value),
-    Identifier(String),
-}
