--- src/ir/ir_nodes/mod.rs
@@ -1,12 +1,0 @@
-#![allow(unused)]
-pub mod assign;
-pub mod binop;
-pub mod block;
-pub mod call;
-pub mod expr;
-pub mod func;
-pub mod stmt;
-pub mod variable;
-pub mod data_type;
-pub mod loops;
-pub mod unaryop;
