--- src/ir/ir_nodes/stmt.rs
@@ -1,8 +1,0 @@
-use super::{assign::IRAssignement, variable::IRVariable, loops::{IRWhile, IRFor}};
-
-pub enum IRStatement {
-    Assignement(IRAssignement),
-    Variable(IRVariable),
-    While(IRWhile),
-    For(IRFor),
-}
