--- src/ir/ir_nodes/unaryop.rs
@@ -1,10 +1,0 @@
-use super::expr::IRExpression;
-
-pub struct IRUnaryOp  {
-    op: UnaryOperator,
-    expr: Box<IRExpression>,
-}
-pub enum UnaryOperator {
-    Negate,
-    Not,
-}
