--- src/vm/events.rs
@@ -1,9 +1,0 @@
-pub enum VMEvent {
-    IllegalOpCode,
-    OutOfBound,
-    StackOverflow,
-    StackUnderflow,
-    DivideByZero,
-    ExitSyscall,
-    AllGood,
-}
