--- src/vm/mod.rs
@@ -1,4 +1,0 @@
-pub mod bytecode;
-pub mod events;
-pub mod virtual_cpu;
-pub mod config;
